requirements.txt
requests
pandas
ta
gspread
google-auth
pyyaml
pandas_ta
